refactor: Move NextAuth configuration to dedicated file
